+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod crypto;
+pub mod ledger;
 pub mod network;
-pub mod licencing;
